Move middlewares out of the router
