+++ st2auth/st2auth/app.py
@@ -15,15 +15,24 @@
 
 import pkg_resources
 
+import jinja2
 from oslo_config import cfg
 import yaml
 
 from st2auth import config as st2auth_config
+import st2common.constants.pack
+import st2common.constants.action
 from st2common import log as logging
-from st2common.router import Router, ErrorHandlingMiddleware
+from st2common.middleware.error_handling import ErrorHandlingMiddleware
+from st2common.middleware.cors import CorsMiddleware
+from st2common.middleware.request_id import RequestIDMiddleware
+from st2common.middleware.logging import LoggingMiddleware
+from st2common.rbac.types import PermissionType
+from st2common.router import Router
 from st2common.util.monkey_patch import monkey_patch
 from st2common.constants.system import VERSION_STRING
 from st2common.service_setup import setup as common_setup
+from st2common.util import isotime
 
 LOG = logging.getLogger(__name__)
 
@@ -49,14 +58,26 @@ def setup_app(config={}):
                      run_migrations=False,
                      config_args=config.get('config_args', None))
 
-    spec_string = pkg_resources.resource_string(__name__, 'controllers/openapi.yaml')
-    spec = yaml.load(spec_string)
+    arguments = {
+        'DEFAULT_PACK_NAME': st2common.constants.pack.DEFAULT_PACK_NAME,
+        'LIVEACTION_STATUSES': st2common.constants.action.LIVEACTION_STATUSES,
+        'PERMISSION_TYPE': PermissionType,
+        'ISO8601_UTC_REGEX': isotime.ISO8601_UTC_REGEX
+    }
 
     router = Router(debug=cfg.CONF.auth.debug)
+
+    spec_template = pkg_resources.resource_string(__name__, 'controllers/openapi.yaml')
+    spec_string = jinja2.Template(spec_template).render(**arguments)
+    spec = yaml.load(spec_string)
+
     router.add_spec(spec)
 
     app = router.as_wsgi
 
+    app = CorsMiddleware(app)
+    app = LoggingMiddleware(app, router)
     app = ErrorHandlingMiddleware(app)
+    app = RequestIDMiddleware(app)
 
     return app
